Module Sesame

Sesame

A small build tool for static blogs and sites with a mission to help reduce site size and promote good accessibility practices.

Building

module Collection : sig ... end

Collections are groups of documents which share the same shape of meta data

module Components : sig ... end

Components contains some useful Txyml.Html element and document geneators for when you want to customise your build process

module Build : sig ... end

Build takes a Collection and produces useful builder functions

module Image : sig ... end

A small library for manipulating images

module Files : sig ... end

A small library for handling files

module Utils : sig ... end

A small collection of useful utility functions

Checking

module Access : sig ... end

Access contains functions for checking whether code meets some accessibility standards

module Checks : sig ... end

The framework that Access is built on, exposed so you can write your own checks. Completely inspired by Alcotest