Module Sesame.Checks

The framework that Access is built on, exposed so you can write your own checks. Completely inspired by Alcotest

module type CHECKABLE = sig ... end

type 'a checkable = (module CHECKABLE with type t = 'a)

val checkable : 'a Fmt.t -> ('a -> bool) -> 'a checkable

Checkable constructor

val md : (Omd.doc -> bool) -> Omd.doc checkable

Predefined markdown checkable

val html : (Tyxml.Html.doc -> bool) -> Tyxml.Html.doc checkable

Predefined Tyxml HTML checkable

val check : ?⁠exit:bool -> string -> 'a checkable -> 'a -> unit

check msg checkable t performs a check on t using checkable described by msg. If you don't want the function to fail if the check fails then pass ~exit:false