Module Image.Transform

type conf = {

quality : int;

rename : string -> string;

files : Fpath.t list;

dst : Fpath.t;

}

A configuration for transformations

val transform : conf:conf -> (t -> t) list -> unit

transform ~conf transforms will load all of the images in conf.files and it will then apply the list of transforms to the images before placing the output files in dir. The rename option lets you specify a function that renames the image, the quality option let's you specify the jpeg quality defaulting to 60