Module Responsive.Images

Generate Responsive Images

type t =

| MaxWidth of int * int * t

| Default of int

type conf = {

root : Fpath.t;

conf : Image.Transform.conf;

}

A slightly extended image conf, you need to also provide the file to which the images are relative too.

val v : alt:string -> conf:conf -> t -> (Fpath.t * [> Html_types.img ] Tyxml_html.elt) list

v ~alt ~conf t applies the responsive image optimisations to all of the images in the transformation configuration, conf, and produces an image srcset for each (it returns an association list with key being the original image's filepath).