Module Sesame__.Collection

module type Meta = sig ... end

module type S = sig ... end

module Make : functor (M : Meta) -> S with type meta = M.t

Make (M) is the recommended way to generate a Collection module -- you provide the Meta module which can use ppx_deriving_yaml to automatically generate the of_yaml and to_yaml functions.