Module Sesame__.Responsive

Tools for Responsive Web Pages

This modules contains different tools for building more responsive web pages. In particular, its biggest selling feature is the responsive images module which can generate <img srcset=...> HTML tags from a source image i.e. it can create the rescaled images, save them somewhere and give back to you the HTML tag to use as you see fit

module Images : sig ... end