Module Sesame__Components

val html : ?⁠lang:string -> title:string -> description:string -> body:[< Html_types.flow5 PCDATA ] Tyxml.Html.elt list -> Tyxml.Html.doc

html ~title ~description ~body will construct an HTML document with the meta field set to description the title field set to title and the body set to body. You can also provide an optional language (default 'en')