Module Sesame__Transformer

Transformer

This Transformer module holds different markdown to markdown transformations (if you can think of more please add them). This is a nice way to extend Omd's capabilities in a simple, non-intrusive way.

The Toc module does what it says on the tin. It takes markdown document, extracts the headers and produces a table of contents based on those headers that you can add and style as you see fit. You'll need to see the generated toc or have a look at the source code to find what classes to style in your CSS... Or you could use the low-level tools and write your own to_html function

module Toc : sig ... end

Images

The images transformer uses the Sesame.Responsive module under-the-hood to provide easy-to-add responsive image handling for markdown documents. It comes with the fairly big caveat that the images in the markdown document should be specified relative to that document as it made path resolution much easier.

It will rewrite any occurence of ![text](./image/path) into a responsive image HTML tag with srcset generated by the user supplying a Sesame.Responsive.Images.t

module Image : sig ... end