Module Sesame__Transformer.Image

type t

Some configuration for the images being transformed in a markdown file

val v : quality:int -> path:Fpath.t -> dst:Fpath.t -> Sesame.Responsive.Images.t -> t

v ~quality ~dst responsive is a constructor for this transform's configuration value

val transform : t -> Omd.doc -> Omd.doc

transform t doc transforms all of the images in doc a specified by the configuration