Module Sesame__.Collection

Collections

Collections are at the very heart of Sesame and was the first thing Sesame could do. A Collection is simply a set of files that share the same metadata structure. This is expressed in the "Jekyll Format", a header (frontmatter) of yaml followed by markdown.

Once you have specified the shape of the metadata using types, Sesame can more or less handle the rest until you want to customise the output which you most likely will!

module type Meta = sig ... end

Make a Collection

This module creates a collection from a file

module Make : functor (M : Meta) -> sig ... end

Make HTML Output

The HTML functor generates a string of HTML based on some metadata (i.e. a collection)

module Html : functor (M : Meta) -> sig ... end